Application 实现 Launcher 和 Daemon
