+++ buession-springboot-boot/src/main/java/com/buession/springboot/boot/application/Launcher.java
@@ -1,0 +1,97 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more contributor license agreements.
+ * See the NOTICE file distributed with this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to you under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is
+ * distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and limitations under the License.
+ *
+ * =========================================================================================================
+ *
+ * This software consists of voluntary contributions made by many individuals on behalf of the
+ * Apache Software Foundation. For more information on the Apache Software Foundation, please see
+ * <http://www.apache.org/>.
+ *
+ * +-------------------------------------------------------------------------------------------------------+
+ * | License: http://www.apache.org/licenses/LICENSE-2.0.txt 										       |
+ * | Author: Yong.Teng <webmaster@buession.com> 													       |
+ * | Copyright @ 2013-2023 Buession.com Inc.														       |
+ * +-------------------------------------------------------------------------------------------------------+
+ */
+package com.buession.springboot.boot.application;
+
+/**
+ * 应用启动器
+ *
+ * @author Yong.Teng
+ * @since 2.2.0
+ */
+public interface Launcher {
+
+	/**
+	 * 启动应用程序
+	 *
+	 * @param args
+	 * 		启动参数
+	 */
+	void startup(final String[] args);
+
+	/**
+	 * 启动应用程序
+	 *
+	 * @param clazz
+	 * 		启动应用程序类
+	 * @param args
+	 * 		启动参数
+	 */
+	void startup(final Class<? extends Application> clazz, final String[] args);
+
+	/**
+	 * 启动应用程序
+	 *
+	 * @param args
+	 * 		启动参数
+	 */
+	default void start(final String[] args){
+		startup(args);
+	}
+
+	/**
+	 * 启动应用程序
+	 *
+	 * @param clazz
+	 * 		启动应用程序类
+	 * @param args
+	 * 		启动参数
+	 */
+	default void start(final Class<? extends Application> clazz, final String[] args){
+		startup(clazz, args);
+	}
+
+	/**
+	 * 启动应用程序
+	 *
+	 * @param args
+	 * 		启动参数
+	 */
+	default void launch(final String[] args){
+		startup(args);
+	}
+
+	/**
+	 * 启动应用程序
+	 *
+	 * @param clazz
+	 * 		启动应用程序类
+	 * @param args
+	 * 		启动参数
+	 */
+	default void launch(final Class<? extends Application> clazz, final String[] args){
+		startup(clazz, args);
+	}
+
+}
